format spring-4.3.x-scenario
