+++ test/plugin/pom.xml
@@ -196,8 +196,8 @@
                         <sourceDirectory>scenarios/solrj-7.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/spring-3.0.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/spring-3.1.x-scenario</sourceDirectory>
-<!--                        <sourceDirectory>scenarios/spring-4.1.x-scenario</sourceDirectory>-->
-<!--                        <sourceDirectory>scenarios/spring-4.3.x-scenario</sourceDirectory>-->
+                        <sourceDirectory>scenarios/spring-4.1.x-scenario</sourceDirectory>
+                        <sourceDirectory>scenarios/spring-4.3.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/spring-async-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/spring-cloud-feign-1.1.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/spring-cloud-feign-1.2.x-scenario</sourceDirectory>
