--- test/plugin/scenarios/spring-4.3.x-scenario/src/main/java/test/org/apache/skywalking/apm/testcase/implinterface/TestCaseController.java
@@ -1,36 +1,0 @@
-/*
- *
- *   Licensed to the Apache Software Foundation (ASF) under one or more
- *   contributor license agreements.  See the NOTICE file distributed with
- *   this work for additional information regarding copyright ownership.
- *   The ASF licenses this file to You under the Apache License, Version 2.0
- *    (the "License"); you may not use this file except in compliance with
- *    the License.  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- *
- */
-
-package test.org.apache.skywalking.apm.testcase.implinterface;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestCaseController implements TestCaseInterface {
-
-    @Override
-    public String implRequestMappingAnnotationTestCase() {
-        return "implRequestMappingAnnotationTestCase";
-    }
-
-    @Override
-    public String implRestAnnotationTestCase() {
-        return "implRestAnnotationTestCase";
-    }
-}
